Fix the playlist generator to allow for spaces
Also support files that exist in other directories.

This also attempts to fix the issue of completely and fully stopping and flushing
a player before attempting to queue any new music to it.

IE, if a player is "paused" and it's "now playing" is a Pandora station,
there's no way to get it to leave that pandora station.  Kinda a bother.
Still needs work.